fix: dependency build
ANIMA .
