+++ .gitignore
@@ -5,6 +5,5 @@
 /bin
 /dist
 /build
-/lib
 /tmp
 node_modules
