+++ design-tokens.json
@@ -1,0 +1,6 @@
+{
+  "colorPrimary": {
+    "$value": "#000000",
+    "$type": "color"
+  }
+}
