+++ src/constants/design-tokens.ts
@@ -1,0 +1,1 @@
+export const TOKEN_COLOR_TYPE = 'color' as const;
