+++ src/constants/index.ts
@@ -1,6 +1,5 @@
 import { API_URL } from './api';
 
-export const TOKEN_COLOR_TYPE = 'color' as const;
 export const TMP_DIR = './tmp';
 export const STORYBOOK_SERVICE_BASE_URL = `${API_URL}/services/s2f`;
 export { API_URL };
