+++ src/converters/utils.ts
@@ -1,4 +1,4 @@
-import { TOKEN_COLOR_TYPE } from "../constants";
+import { TOKEN_COLOR_TYPE } from "../constants/design-tokens";
 import type { TokenValue } from "../constants/types";
 
 export const formatColorToTokenValue = (currentValue: string): TokenValue => ({
