+++ theme.js
@@ -1,0 +1,5 @@
+module.exports = {
+  token: {
+    colorPrimary: "#000000",
+  },
+}
